backend/requirements.txt
fastapi
uvicorn[standard]
googletrans==4.0.0-rc1
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib


